requirements.txt
python-dotenv
kafka-python
feast
mlflow
dvc
prometheus-client
requests
MetaTrader5
pandas
numpy
scikit-learn
joblib
torch
transformers
SQLAlchemy
loguru
pytest
pytest-cov
flake8
isort
black
mypy
bandit
safety
pip-audit
